requirements.txt
streamlit-nightly
plotly
authlib